+++ core/lib/gtm.ts
@@ -1,4 +1,4 @@
-export const GTM_ID = process.env.NEXT_PUBLIC_GTM_ID || 'GTM-KGQLWDKB';
+export const GTM_ID = process.env.NEXT_PUBLIC_GTM_ID || '';
 
 declare global {
   interface Window {
